Animations and tweening for cat attack
crashes if there are randomly no cats in lane 1 or 2. Do we really need the cat objects to be hidden in lanes arrays?

Please make the construction of a cat to add a list of animations to its sprite, and then allow the card playing action to fire their animations and run that movement thing chainto tween etc.